+++ log/src/Record.cc
@@ -22,6 +22,7 @@
 
 #include <ignition/common/Console.hh>
 #include <ignition/transport/Node.hh>
+#include <ignition/transport/Discovery.hh>
 
 #include "ignition/transport/log/Record.hh"
 #include "ignition/transport/log/Log.hh"
@@ -40,8 +41,28 @@ class ignition::transport::log::RecordPrivate
           const std::string &_msgData,
           const transport::MessageInfo &_info);
 
+  /// \brief Callback that listens for newly advertised topics
+  public: void OnAdvertisement(const Publisher &_publisher);
+
+  /// \sa Record::AddTopic(const std::string&)
+  public: RecordError AddTopic(const std::string &_topic);
+
+  /// \sa Record::AddTopic(const std::regex&)
+  public: int AddTopic(const std::regex &_pattern);
+
   /// \brief log file or nullptr if not recording
   public: std::unique_ptr<Log> logFile;
+
+  /// \brief A set of topic patterns that we want to subscribe to
+  public: std::vector<std::regex> patterns;
+
+  /// \brief A set of topic names that we have already subscribed to. When new
+  /// publishers advertise topics that we are already subscribed to, our
+  /// OnAdvertisement callback can just ignore it.
+  public: std::set<std::string> alreadySubscribed;
+
+  /// \brief mutex for thread safety when evaluating newly advertised topics
+  public: std::mutex topicMutex;
 
   /// \brief Value added to steady_clock giving time in UTC
   public: std::chrono::nanoseconds wallMinusMono;
@@ -50,10 +71,13 @@ class ignition::transport::log::RecordPrivate
   public: std::mutex logFileMutex;
 
   /// \brief node used to create subscriptions
-  public: ignition::transport::Node node;
+  public: Node node;
 
   /// \brief callback used on every subscriber
-  public: transport::RawCallback rawCallback;
+  public: RawCallback rawCallback;
+
+  /// \brief Object for discovering new publishers as they advertise themselves
+  public: std::unique_ptr<MsgDiscovery> discovery;
 };
 
 //////////////////////////////////////////////////
@@ -76,6 +100,10 @@ void RecordPrivate::OnMessageReceived(
   igndbg << "RX'" << _info.Topic() << "'[" << _info.Type() << "]\n";
 
   std::lock_guard<std::mutex> lock(this->logFileMutex);
+
+  // Note: this->logFile will only be a nullptr before Start() has been called
+  // or after Stop() has been called. If it is a nullptr, then we are not
+  // recording anything yet, so we can just skip inserting the message.
   if (this->logFile && !this->logFile->InsertMessage(
         timeRX,
         _info.Topic(),
@@ -85,6 +113,76 @@ void RecordPrivate::OnMessageReceived(
   {
     ignwarn << "Failed to insert message into log file\n";
   }
+}
+
+//////////////////////////////////////////////////
+void RecordPrivate::OnAdvertisement(const Publisher &_publisher)
+{
+  std::string partition;
+  std::string topic;
+
+  TopicUtils::DecomposeFullyQualifiedTopic(
+        _publisher.Topic(), partition, topic);
+
+  // If the advertised partition does not match ours, ignore this advertisement
+  if (this->node.Options().Partition() != partition)
+    return;
+
+  // If we are already subscribed to the topic, ignore this advertisement
+  if (this->alreadySubscribed.find(topic) != this->alreadySubscribed.end())
+    return;
+
+  for (const std::regex &pattern : this->patterns)
+  {
+    if (std::regex_match(topic, pattern))
+    {
+      this->AddTopic(topic);
+    }
+  }
+}
+
+//////////////////////////////////////////////////
+RecordError RecordPrivate::AddTopic(const std::string &_topic)
+{
+  igndbg << "Recording [" << _topic << "]\n";
+  // Subscribe to the topic whether it exists or not
+  if (!this->node.RawSubscribe(_topic, this->rawCallback))
+  {
+    ignerr << "Failed to subscribe to [" << _topic << "]\n";
+    return RecordError::FAILED_TO_SUBSCRIBE;
+  }
+
+  this->alreadySubscribed.insert(_topic);
+
+  return RecordError::NO_ERROR;
+}
+
+//////////////////////////////////////////////////
+int RecordPrivate::AddTopic(const std::regex &_topic)
+{
+  int numSubscriptions = 0;
+  std::vector<std::string> allTopics;
+  this->node.TopicList(allTopics);
+  for (auto topic : allTopics)
+  {
+    if (std::regex_match(topic, _topic))
+    {
+      // Subscribe to the topic
+      if (this->AddTopic(topic) == RecordError::FAILED_TO_SUBSCRIBE)
+      {
+        return static_cast<int>(RecordError::FAILED_TO_SUBSCRIBE);
+      }
+      ++numSubscriptions;
+    }
+    else
+    {
+      igndbg << "Not recording " << topic << "\n";
+    }
+  }
+
+  this->patterns.push_back(_topic);
+
+  return numSubscriptions;
 }
 
 //////////////////////////////////////////////////
@@ -103,6 +201,18 @@ Record::Record()
   {
     this->dataPtr->OnMessageReceived(_data, _info);
   };
+
+  Uuid uuid;
+  this->dataPtr->discovery = std::unique_ptr<MsgDiscovery>(
+        new MsgDiscovery(uuid.ToString(), NodeShared::kMsgDiscPort));
+
+  DiscoveryCallback<Publisher> cb = [this](const Publisher &_publisher)
+  {
+    this->dataPtr->OnAdvertisement(_publisher);
+  };
+
+  this->dataPtr->discovery->ConnectionsCb(cb);
+  this->dataPtr->discovery->Start();
 }
 
 //////////////////////////////////////////////////
@@ -150,40 +260,11 @@ void Record::Stop()
 //////////////////////////////////////////////////
 RecordError Record::AddTopic(const std::string &_topic)
 {
-  igndbg << "Recording [" << _topic << "]\n";
-  // Subscribe to the topic whether it exists or not
-  if (!this->dataPtr->node.RawSubscribe(_topic, this->dataPtr->rawCallback))
-  {
-    ignerr << "Failed to subscribe to [" << _topic << "]\n";
-    return RecordError::FAILED_TO_SUBSCRIBE;
-  }
-  return RecordError::NO_ERROR;
+  return this->dataPtr->AddTopic(_topic);
 }
 
 //////////////////////////////////////////////////
 int Record::AddTopic(const std::regex &_topic)
 {
-  int numSubscriptions = 0;
-  std::vector<std::string> allTopics;
-  this->dataPtr->node.TopicList(allTopics);
-  for (auto topic : allTopics)
-  {
-    if (std::regex_match(topic, _topic))
-    {
-      igndbg << "Recording " << topic << "\n";
-      // Subscribe to the topic
-      if (!this->dataPtr->node.RawSubscribe(topic, this->dataPtr->rawCallback))
-      {
-        ignerr << "Failed to subscribe to [" << topic << "]\n";
-        return static_cast<int>(RecordError::FAILED_TO_SUBSCRIBE);
-      }
-      ++numSubscriptions;
-    }
-    else
-    {
-      igndbg << "Not recording " << topic << "\n";
-    }
-  }
-  // TODO store pattern and attempt to subscribe to it when discovered
-  return numSubscriptions;
+  return this->dataPtr->AddTopic(_topic);
 }