+++ Changelog.md
@@ -1,3 +1,8 @@
+## Ignition Transport 4.0.0
+
+1. Added publishing and receiving messages as raw bytes
+    * [Pull request 251](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/251)
+
 ## Ignition Transport 3.0.0
 
 1. Added optional message throttling when publishing messages.